Huge refactor; added config in json

+++ src/ray.h
@@ -7,18 +7,18 @@ namespace raytracer {
 class ray {
    public:
     ray() {}
-    ray(point3 orig, vec3 dir, double t = -1.0) : orig_(orig), dir_(dir), t_(t) {}
+    ray(point3 orig, vec3 dir, float t = -1.0) : orig_(orig), dir_(dir), t_(t) {}
 
     const point3& origin() const { return orig_; }
     const vec3& direction() const { return dir_; }
-    double& t_distance() { return t_; }
+    float& t_distance() { return t_; }
 
-    point3 at(double t) const { return orig_ + dir_ * t; }
+    point3 at(float t) const { return orig_ + dir_ * t; }
 
    private:
     point3 orig_;
     vec3 dir_;
-    double t_;  // t when hit
+    float t_;  // t when hit
 };
 
 }  // namespace raytracer
